2. Интеграционные тесты для проект job4j_cars [#4746]

+++ src/main/java/ru/job4j/cars/repository/CarRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface CarRepository {
     List<Car> getAll();
 
-    Optional<Car> findCarById(int id);
+    Optional<Car> findCarById(Long id);
 
     Car saveCar(Car car);
 }
